+++ crates/gmsol/src/cli/user.rs
@@ -1,5 +1,5 @@
 use gmsol::{
-    store::user::UserOps, types::user::ReferralCode, utils::instruction::InstructionSerialization,
+    store::user::UserOps, types::user::ReferralCodeV2, utils::instruction::InstructionSerialization,
 };
 use gmsol_solana_utils::bundle_builder::BundleOptions;
 use solana_sdk::pubkey::Pubkey;
@@ -20,6 +20,10 @@ enum Command {
     InitReferralCode { code: String },
     /// Transfer Referral Code.
     TransferReferralCode { receiver: Pubkey },
+    /// Cancel referral code transfer.
+    CancelReferralCodeTransfer,
+    /// Accept referral code transfer.
+    AcceptReferralCode { code: String },
     /// Set Referrer.
     SetReferrer { code: String },
 }
@@ -44,14 +48,22 @@ impl Args {
                 .prepare_user(store)?
                 .into_bundle_with_options(options)?,
             Command::InitReferralCode { code } => client
-                .initialize_referral_code(store, ReferralCode::decode(code)?)?
+                .initialize_referral_code(store, ReferralCodeV2::decode(code)?)?
                 .into_bundle_with_options(options)?,
             Command::TransferReferralCode { receiver } => client
                 .transfer_referral_code(store, receiver, None)
                 .await?
                 .into_bundle_with_options(options)?,
+            Command::CancelReferralCodeTransfer => client
+                .cancel_referral_code_transfer(store, None)
+                .await?
+                .into_bundle_with_options(options)?,
+            Command::AcceptReferralCode { code } => client
+                .accept_referral_code(store, ReferralCodeV2::decode(code)?, None)
+                .await?
+                .into_bundle_with_options(options)?,
             Command::SetReferrer { code } => client
-                .set_referrer(store, ReferralCode::decode(code)?, None)
+                .set_referrer(store, ReferralCodeV2::decode(code)?, None)
                 .await?
                 .into_bundle_with_options(options)?,
         };